+++ tests/suite/test_suite.py
@@ -9,7 +9,6 @@ import election
 import code_update
 import membership
 import governance_history
-import replacement
 
 from inspect import signature, Parameter
 
@@ -90,9 +89,8 @@ all_tests_suite = [
     # memberclient:
     memberclient.test_missing_signature_header,
     memberclient.test_corrupted_signature,
-    # replacement
-    replacement.test_node_replacement,
     # reconfiguration:
+    reconfiguration.test_node_replacement,
     reconfiguration.test_add_node,
     reconfiguration.test_add_node_on_other_curve,
     reconfiguration.test_add_node_from_backup,
